major update in utils module to use standard np.asarray methods
